+++ Samples~/Scripts/AdaptExamples/AdaptPropRange.cs
@@ -7,7 +7,16 @@ namespace SaintsField.Samples.Scripts.AdaptExamples
         [Adapt(EUnit.Percent)] public float percentF;
         [Adapt(EUnit.Percent)] public int percentI;
         [PropRange(0f, 1f), Adapt(EUnit.Percent), OverlayRichLabel("<color=gray>%", end: true), BelowRichLabel("$" + nameof(percentRange)), BelowButton("$" + nameof(ExternalPumpValue), "Pump")] public float percentRange;
-        [PropRange(0f, 1f, step: 0.05f), Adapt(EUnit.Percent), OverlayRichLabel("<color=gray>%", end: true), BelowRichLabel("$" + nameof(stepRange)), BelowButton("$" + nameof(ExternalPumpValueStep), "Pump"), BelowButton("$" + nameof(SetValue))] public float stepRange;
+        [
+            PropRange(0f, 1f, step: 0.05f),
+            Adapt(EUnit.Percent),
+            OverlayRichLabel("<color=gray>%", end: true),
+            BelowRichLabel("$" + nameof(DisplayActualValue)),
+            // BelowButton("$" + nameof(ExternalPumpValueStep), "Pump"),
+            // BelowButton("$" + nameof(SetValue))
+        ] public float stepRange;
+
+        private string DisplayActualValue(float av) => $"<color=gray>Actual Value: {av}";
 
         private void ExternalPumpValue(float curValue)
         {
@@ -21,11 +30,13 @@ namespace SaintsField.Samples.Scripts.AdaptExamples
             stepRange = Mathf.FloorToInt((curValue + 0.1f) / 0.1f) % 10 * 0.1f + 0.001f;
         }
 
+        [Space]
         public float v;
 
         private void SetValue()
         {
             stepRange = v;
         }
+
     }
 }