+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.33.0 ##
 
 1.  Add `Adapt` which can work with `PropRange`
-2.  Add `PropRange` support for `Auto Validator`
+2.  Add `PropRange` support for `Auto Validator`9
 3.  UI Toolkit: fix `OverlayRichLabel(end: true)` incorrect overlap
 4.  IMGUI: update `Table` on every draw [#157](https://github.com/TylerTemp/SaintsField/issues/157)